requirements.txt
pandas==2.1.0
numpy==1.24.3
pillow==10.0.0
requests==2.31.0
opencv-python==4.8.0.76
scikit-image==0.21.0
imagehash==4.3.1
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.2.1
selenium==4.12.0
webdriver-manager==4.0.0
scikit-learn==1.3.0
scipy==1.11.2
torch==2.0.1
torchvision==0.15.2
timm==0.9.7  
aiohttp==3.8.5
asyncio==3.4.3
tqdm==4.66.1
redis==5.0.0
diskcache==5.6.3
python-dotenv==1.0.0
pyarrow==13.0.0 